Add files via upload
Due to their large size, the raw sampling data recorded in RNDIS and ECM modes could not be uploaded directly to this repository.

🔗 You can access the data files here: https://drive.google.com/drive/folders/12i0FZn4AJhJjMkElnjOiTy2ANTg5OhjJ?usp=sharing

These files are required for the time analysis and sampling performance comparison conducted in the .py script. You can run the provided analysis code with these recordings to evaluate and compare the sustainable data rates of RNDIS and ECM protocols.